lab 2 (part)
Added tab navigation and made lab 1 and lab 2 separate
(still has not resolved checkbox issue)
