+++ vasilevIgor/components/navigations/tabNavigation.js
@@ -1,0 +1,38 @@
+import * as React from 'react';
+import { Text, View } from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import Lab1 from "../screens/lab1";
+import Lab2 from "../screens/lab2";
+
+const Tab = createBottomTabNavigator();
+
+const tabNavigation = () => {
+
+    return (
+        <NavigationContainer>
+          <Tab.Navigator
+            screenOptions={({ route }) => ({
+              tabBarIcon: ({ focused, color, size }) => {
+                let iconName;
+    
+                if (route.name === 'Home') {
+                  iconName = focused
+                    ? 'ios-information-circle'
+                    : 'ios-information-circle-outline';
+                } else if (route.name === 'Settings') {
+                  iconName = focused ? 'ios-list' : 'ios-list-outline';
+                }
+    
+                return <Ionicons name={iconName} size={size} color={color} />;
+              },
+              tabBarActiveTintColor: 'tomato',
+              tabBarInactiveTintColor: 'gray',
+            })}
+          >
+            <Tab.Screen name="lab1" component={Lab1} />
+            <Tab.Screen name="lab2" component={Lab2} />
+          </Tab.Navigator>
+        </NavigationContainer>
+      );
+    }
