+++ vasilevIgor/components/screens/lab1/noteItem.js
