+++ vasilevIgor/package.json
@@ -10,6 +10,8 @@
   },
   "dependencies": {
     "@react-native-community/checkbox": "^0.5.15",
+    "@react-navigation/bottom-tabs": "^6.5.7",
+    "@react-navigation/native": "^6.1.6",
     "expo": "~47.0.12",
     "expo-status-bar": "~1.4.2",
     "react": "18.1.0",
